work around some harmful signedness in MuPDF and fix two crashes (fixes issue 1289)

+++ mupdf/fitz/stm_read.c
@@ -110,7 +110,8 @@ fz_readall(fz_buffer **bufp, fz_stream *stm, int initial)
 		if (buf->len == buf->cap)
 			fz_growbuffer(buf);
 
-		if (buf->len > initial * 200)
+		/* http://code.google.com/p/sumatrapdf/issues/detail?id=1289 */
+		if ((size_t)buf->len > (size_t)initial * 200)
 		{
 			fz_dropbuffer(buf);
 			return fz_throw("compression bomb detected");
